third_party/blink/web_tests/http/tests/security/mixedContent/strict-mode-via-pref-image-blocked.https-expected.txt

CONSOLE ERROR: line 13: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/strict-mode-via-pref-image-blocked.https.html' was loaded over HTTPS, but requested an insecure image 'http://example.test:8080/security/resources/compass.jpg'. This request has been blocked; the content must be served over HTTPS.
This test passes if the image below is treated as blockable content.
